extract teacher metrics into screen, fix back button
