+++ tutor/resources/styles/tutor.scss
@@ -22,7 +22,6 @@
 @import './components/media-preview';
 @import './components/my-courses';
 @import './components/tutor-dialog';
-@import './components/task-teacher-review/index';
 @import './components/plan-stats';
 @import './components/multi-select';
 @import './components/registration/index';
