+++ tutor/src/routes.js
@@ -49,7 +49,7 @@ const getRoutes = (router) => {
           ],
         }, {
           path: 'metrics/:id', name: 'reviewTask',
-          renderer: r(() => import('./components/task-teacher-review')) },
+          renderer: r(() => import('./screens/teacher-review-metrics')) },
         { path: 'task/:id', name: 'viewTask',
           renderer: r(() => import('./screens/task'), 'Assignment'),
           routes: [
